feat: slug in city

+++ backend/src/modules/cities/use-cases/CityDeleteUseCase.ts
@@ -7,10 +7,21 @@ export class CityDeleteUseCase {
     ) {}
 
     async execute(id: string): Promise<void> {
-        const city = await this.cityRepository.findUnique(id);
+        const city = await this.cityRepository.findUnique(id)
 
-        if (!city) throw new ServerError("City not found", 404);
+        if (!city) throw new ServerError("City not found", 404)
+    
 
-        await this.cityRepository.delete(id);
+        const relations = await this.cityRepository.findPlacesAndEventsById(id)
+
+        if (!relations) throw new ServerError("City relations not found", 404)
+
+        const hasPlaces = relations.places && relations.places.length > 0
+        const hasEvents = relations.events && relations.events.length > 0
+
+        if (hasPlaces || hasEvents) throw new ServerError("Cannot delete city with associated places or events", 400)
+    
+
+        await this.cityRepository.delete(id)
     }
 }
